+++ pjrt/cpu/static/static.go
@@ -5,6 +5,8 @@
 //	import _ "github.com/gomlx/gopjrt/pjrt/static"
 //
 // And calls to pjrt.GetPlugin("cpu") will return the statically linked one.
+//
+// If you use this, don't use prjt/cpu/dynamic package. Only one can be set.
 package static
 
 // #cgo LDFLAGS: -lpjrt_c_api_cpu_static -lstdc++ -lm
@@ -23,10 +25,10 @@ import (
 func init() {
 	pjrtAPI := uintptr(unsafe.Pointer(C.GetPjrtApi()))
 	if pjrtAPI == 0 {
-		klog.Fatal("Failed to get PJRT API pointer when initializing static PJRT (github.com/gomlx/gopjrt/pjrt/static).")
+		klog.Fatal("Failed to get PJRT API pointer when initializing statically preloaded PJRT (github.com/gomlx/gopjrt/pjrt/cpu/static).")
 	}
 	err := pjrt.RegisterPreloadedPlugin("cpu", pjrtAPI)
 	if err != nil {
-		klog.Fatalf("Failed to register static PJRT plugin for CPU (github.com/gomlx/gopjrt/pjrt/static): %+v", err)
+		klog.Fatalf("Failed to register statically preloaded PJRT plugin for CPU (github.com/gomlx/gopjrt/pjrt/cpu/static): %+v", err)
 	}
 }