+++ docs/CHANGELOG.md
@@ -1,7 +1,11 @@
 # Next
 
-* Added optional `github.com/gomlx/gopjrt/pjrt/static` that statically links the PJRT CPU plugin. 
-  It includes the corresponding C BUILD rule to build the static library (`libpjrt_c_api_cpu.a`)
+* Optional preloading CPU PJRT plugin:
+  * `github.com/gomlx/gopjrt/pjrt/cpu/static` that statically links the PJRT CPU plugin: easy to deploy binary. 
+    It includes the corresponding C BUILD rule to build the static library (`libpjrt_c_api_cpu_static.a`)
+  * `github.com/gomlx/gopjrt/pjrt/cpu/dynamic` that dynamically links (and preloads) the PJRT CPU plugin.
+* `pjrt_c_api_cpu.so` now compiled directly from `gopjrt`, and doesn't require cloning `xla` separately. It will
+  be distributed in the same `tar.gz` file.
 
 # v0.4.8 - 2024-11-19